# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Pourier API'
  description: "API REST pour la plateforme de vente de photos Pourier. Cette API permet aux utilisateurs d'acheter des photos, aux photographes de vendre leurs créations, et aux administrateurs de gérer la plateforme."
  contact:
    name: 'Pourier Support'
    email: contact@pourier.com
  license:
    name: Proprietary
    url: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Serveur API Pourier'
paths:
...
  /api/admin/users:
    get:
      tags:
        - 'Admin - Users'
      summary: 'Get all users with filtering'
      description: 'Retrieve all users with advanced filtering options by account type, active status, and search term. Requires admin role.'
      operationId: getAdminUsers
      parameters:
        -
          name: account_type
          in: query
          description: 'Filter by account type'
          required: false
          schema:
            type: string
            enum:
              - buyer
              - photographer
              - admin
            example: buyer
        -
          name: is_active
          in: query
          description: 'Filter by active status'
          required: false
          schema:
            type: boolean
            example: true
        -
          name: search
          in: query
          description: 'Search by name or email'
          required: false
          schema:
            type: string
            example: Jean
        -
          name: per_page
          in: query
          description: 'Number of items per page'
          required: false
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        200:
          description: 'Users retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      data:
                        type: array
                        items:
                          type: object
                      total:
                        type: integer
                        example: 1250
                    type: object
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
        403:
          description: 'Forbidden - Admin role required'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Accès non autorisé'
                type: object
      security:
        -
          bearerAuth: []
  '/api/admin/users/{user}':
    get:
      tags:
        - 'Admin - Users'
      summary: 'Get user details'
      description: 'Retrieve detailed information about a specific user. For photographers, includes profile and recent photos. Requires admin role.'
      operationId: getAdminUser
      parameters:
        -
          name: user
          in: path
          description: 'User UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      responses:
        200:
          description: 'User details retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
        403:
          description: 'Forbidden - Admin role required'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Accès non autorisé'
                type: object
        404:
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Utilisateur introuvable'
                type: object
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Admin - Users'
      summary: 'Delete user account'
      description: 'Soft delete a user account. Admin accounts cannot be deleted. Requires admin role.'
      operationId: deleteAdminUser
      parameters:
        -
          name: user
          in: path
          description: 'User UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      responses:
        200:
          description: 'User account deleted successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Compte supprimé avec succès.'
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
        403:
          description: 'Forbidden - Cannot delete admin account or admin role required'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Impossible de supprimer un compte administrateur.'
                type: object
        404:
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Utilisateur introuvable'
                type: object
      security:
        -
          bearerAuth: []
  '/api/admin/users/{user}/suspend':
    put:
      tags:
        - 'Admin - Users'
      summary: 'Suspend user account'
      description: "Suspend a user's account, preventing them from accessing the platform. Requires admin role."
      operationId: suspendAdminUser
      parameters:
        -
          name: user
          in: path
          description: 'User UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      responses:
        200:
          description: 'User account suspended successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Compte suspendu avec succès.'
                type: object
        400:
          description: 'Account already suspended'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Ce compte est déjà suspendu.'
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
        403:
          description: 'Forbidden - Admin role required'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Accès non autorisé'
                type: object
        404:
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Utilisateur introuvable'
                type: object
      security:
        -
          bearerAuth: []
  '/api/admin/users/{user}/activate':
    put:
      tags:
        - 'Admin - Users'
      summary: 'Activate user account'
      description: "Activate a suspended user's account, restoring their access to the platform. Requires admin role."
      operationId: activateAdminUser
      parameters:
        -
          name: user
          in: path
          description: 'User UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      responses:
        200:
          description: 'User account activated successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Compte activé avec succès.'
                type: object
        400:
          description: 'Account already active'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Ce compte est déjà actif.'
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
        403:
          description: 'Forbidden - Admin role required'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Accès non autorisé'
                type: object
        404:
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Utilisateur introuvable'
                type: object
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    UnauthorizedResponse:
      title: 'Unauthorized Response'
      description: 'Réponse pour une authentification requise'
      properties:
        message:
          type: string
          example: 'Non authentifié'
      type: object
...
  responses:
    UnauthorizedResponse:
      description: 'Unauthorized - Authentication required'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter JWT Bearer token obtained from /api/auth/login endpoint'
...
security:
  -
    bearerAuth: []